requirements.txt
flask
gunicorn
tensorflow==2.10.0
keras==2.10.0
numpy==1.23.5
scipy==1.9.3
opencv-python-headless
pillow
